# specs/001-fittrack-mvp/contracts/api.yaml
openapi: 3.1.0
info:
  title: FitTrack Pro MVP API
  version: 1.0.0
  description: REST API for FitTrack Pro fitness tracking application
servers:
  - url: /api
    description: Next.js API Routes
security:
  - clerkAuth: []
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, unitPreference, fitnessLevel]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        weightKg:
          type: number
          nullable: true
        unitPreference:
          type: string
          enum: [metric, imperial]
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        createdAt:
          type: string
          format: date-time
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        weightKg:
          type: number
          minimum: 20
          maximum: 300
        unitPreference:
          type: string
          enum: [metric, imperial]
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced]
    EquipmentProfile:
      type: object
      required: [dumbbells, barbells, kettlebells, bodyweight]
      properties:
        dumbbells:
          type: boolean
        barbells:
          type: boolean
        kettlebells:
          type: boolean
        bodyweight:
          type: boolean
          default: true
...
    Workout:
      type: object
      required: [id, userId, type, startedAt, durationSeconds]
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [treadmill, strength]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        durationSeconds:
          type: integer
        notes:
          type: string
          nullable: true
        synced:
          type: boolean
        localId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    TreadmillWorkout:
      allOf:
        - $ref: '#/components/schemas/Workout'
        - type: object
          required: [distanceKm, avgSpeedKmh]
          properties:
            distanceKm:
              type: number
            avgSpeedKmh:
              type: number
            caloriesBurned:
              type: integer
              nullable: true
    CreateTreadmillWorkoutInput:
      type: object
      required: [startedAt, durationSeconds, distanceKm]
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          minimum: 1
          maximum: 86400
        distanceKm:
          type: number
          minimum: 0.01
          maximum: 100
        notes:
          type: string
          maxLength: 500
        localId:
          type: string
          description: Client-generated ID for offline sync
    StrengthExerciseLog:
      type: object
      required: [exerciseId, exerciseName, sets]
      properties:
        exerciseId:
          type: string
        exerciseName:
          type: string
        sets:
          type: array
          items:
            type: object
            required: [reps, completed]
            properties:
              reps:
                type: integer
                minimum: 1
              weightKg:
                type: number
                minimum: 0
              completed:
                type: boolean
    StrengthWorkout:
      allOf:
        - $ref: '#/components/schemas/Workout'
        - type: object
          required: [exercises]
          properties:
            planId:
              type: string
              nullable: true
            exercises:
              type: array
              items:
                $ref: '#/components/schemas/StrengthExerciseLog'
    CreateStrengthWorkoutInput:
      type: object
      required: [startedAt, durationSeconds, exercises]
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          minimum: 1
        planId:
          type: string
          nullable: true
        exercises:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/StrengthExerciseLog'
        notes:
          type: string
          maxLength: 500
        localId:
          type: string
...
    Exercise:
      type: object
      required: [id, name, description, muscleGroups, equipment, difficulty, instructions]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        muscleGroups:
          type: array
          items:
            type: string
            enum: [chest, back, shoulders, biceps, triceps, core, quadriceps, hamstrings, glutes, calves]
        equipment:
          type: string
          enum: [bodyweight, dumbbells, barbells, kettlebells]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        instructions:
          type: string
        imageUrl:
          type: string
          nullable: true

    # ============ Workout Plans ============
    PlanExercise:
      type: object
      required: [exerciseId, exerciseName, sets, reps, restSeconds]
      properties:
        exerciseId:
          type: string
        exerciseName:
          type: string
        sets:
          type: integer
          minimum: 1
        reps:
          oneOf:
            - type: integer
            - type: string
          description: Number like 12, or range like "8-12", or duration like "30 seconds"
        restSeconds:
          type: integer
          minimum: 0
        notes:
          type: string
    WorkoutPlan:
      type: object
      required: [id, userId, name, goal, durationMinutes, exercises, isAiGenerated]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        goal:
          type: string
          enum: [build_muscle, lose_weight, general_fitness]
        durationMinutes:
          type: integer
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/PlanExercise'
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /user/profile:
    get:
      summary: Get current user profile
      tags: [User]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
    
    patch:
      summary: Update user profile
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /user/equipment:
    get:
      summary: Get user equipment profile
      tags: [User]
      responses:
        '200':
          description: Equipment profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentProfile'
    
    put:
      summary: Update equipment profile
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentProfile'
      responses:
        '200':
          description: Updated equipment profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentProfile'
...
  /workouts/treadmill:
    post:
      summary: Create treadmill workout
      tags: [Workouts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTreadmillWorkoutInput'
      responses:
        '201':
          description: Created workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreadmillWorkout'
        '409':
          description: Conflict - localId already exists
  /workouts/strength:
    post:
      summary: Create strength workout
      tags: [Workouts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrengthWorkoutInput'
      responses:
        '201':
          description: Created workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrengthWorkout'
  /workouts/{id}:
    get:
      summary: Get workout by ID
      tags: [Workouts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workout details
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TreadmillWorkout'
                  - $ref: '#/components/schemas/StrengthWorkout'
        '404':
          description: Workout not found

    delete:
      summary: Delete workout
      tags: [Workouts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
  /exercises:
    get:
      summary: List exercises
      tags: [Exercises]
      parameters:
        - name: equipment
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [bodyweight, dumbbells, barbells, kettlebells]
          style: form
          explode: true
        - name: muscleGroup
          in: query
          schema:
            type: string
            enum: [chest, back, shoulders, biceps, triceps, core, quadriceps, hamstrings, glutes, calves]
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: userEquipmentOnly
          in: query
          schema:
            type: boolean
            default: false
          description: Filter by user's equipment profile
      responses:
        '200':
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
...
  /plans/{id}:
    get:
      summary: Get plan by ID
      tags: [Plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutPlan'
        '404':
          description: Not found

    delete:
      summary: Delete saved plan
      tags: [Plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...